add1

+++ js/task-3.js
@@ -1,6 +1,6 @@
 'use strict';
 function getElementWidth(content, padding, border) {
-    const boxWidth = Number.parseInt(content) + Number.parseInt(padding) * 2 + Number.parseInt(border) * 2;
+    const boxWidth = Number.parseFloat(content) + Number.parseFloat(padding) * 2 + Number.parseFloat(border) * 2;
     return boxWidth;
 }
 console.log(getElementWidth("50px", "8px", "4px")); // 74
